requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
drf-spectacular
django-cors-headers
